adding warpgate and switch to crystalc.gd
Trying to get the jump out portal animation to work.

This still crash when giving the warp flag to crystalc-warp. Giving a warp flag to a level with the portal does seems to work.
The spawn position in the checkpoint "should be" the one you land on and then the warpgate or the level-part is where you jump from so I need to be able to link the two or some shit idk.

When setting one too far away it doesn't go as far as the actual checkpoint is tho.

If set on a checkpoint of a level that's loaded at the same time as a portal (ex lava-tube-end) the portal sound plays but jak doesn't spawn there and the camera is stuck around the spawn point (softlock but no crash)

Probably need to add it with a x-part.gc thing. Gotta figure out how that works.